[Fix] Remove incorrect use of VarNR
Summary:
I had put in VarNR for a case that never needed ref-counting, thinking it was a
synonym for "Variant without refcounting, and without a destructor". But it
doesnt have any of the integer constructors, and it doesnt support
VarNR(CVarRef). Change it to Variant - this case really doesnt need optimizing;
if the lhs of an -> operator in rval context isnt an object, we're going to
fatal anyway...

Test Plan:
fast_tests slow_tests

DiffCamp Revision: 197427
Reviewed By: qigao
Reviewers: qigao
CC: qigao, hphp-diffs@lists
Tasks:
459860: varNR assertion failure

Revert Plan:
OK