[php53] implemented GOTO
Summary:
GOTO is needed for "yield" to work, so I will have 3 or 4 batches of changes
implemented in this order:

  - GOTO (this diff)
  - anonymous function / closure / continuation
  - namespace
  - type hints

There is only one problem that's inconsistent with PHP: I'm not allowing
cross-try-catch gotos, which isn't allowed in C++. Even in PHP it's confusing,
whether it means having a "try" in effect or not, when going into a try block. I
find allowing it really bad or strange. If we feel we should support it, we can
come up with some code transformation in generated C++ code to make it work.
Otherwise, that seems to be a bad coding style on top of an already-bad GOTO
coding style.

Test Plan:
fast_tests, slow_tests

DiffCamp Revision: 197446
Reviewed By: mwilliams
Reviewers: mwilliams
CC: hzhao, achao, mwilliams, hphp-diffs@lists
Tasks:
293021: PHP 5.3 and beyond

Revert Plan:
OK