[Perf/stability] Initialize m_hasOverride
Summary:
A new flag was not being cleared, causing overly conservative code generation
if it happened to be initialized true (note that before the flag was added, we
assumed it was true, so this was still better than before, and correct; but code
generation was not repeatable).

Also fix a loop over a vector that might change during the iteration.

Test Plan:
fast_tests, slow_tests, build www-dev multiple times and compare generated
files.

DiffCamp Revision: 197470
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK