Fix possibly wrong LINE macro call syntax
Summary:
Patch from hyves:

In case an expression is passed as the second argument of the LINE
macro, and that expression involves calling a template with 2
arguments, the preprocessor can get confused since it doesn't recognize
the template<a,b> syntax.

For example compiling the following code:
<?php
class A {
	function foo() {
		throw "ERROR";
	}
	function test() {
		$this->i %= foo();
	}
}
?>

generates the C++ statement:
 LINE(7,GET_THIS_DOT()o_assign_op<void,272>(...));

where the preprocessor interprets the comma in the template parameters
as separating a second macro parameter from a third one, and thus
giving an error.

This patch fixes the problem by using parentheses to specify the
expressions.

Test Plan:
make fast_tests
make slow_tests

DiffCamp Revision: 197693
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK