[Fix] Prevent incorrect error from array_intersect_key
Summary:
array_intersect_key was recently changed to raise an error if either of the
first two params is not an array. But php doesnt raise an error if the first
param is an empty array, and the second is not an array.

Test Plan:
fast_tests, slow_tests,
flib/pages/fans/friends/__tests__/PageFansFriendsFetcherTestCase.php

DiffCamp Revision: 198247
Reviewed By: hzhao
Reviewers: hzhao
Commenters: qixin
CC: hzhao, qixin, mwilliams, hphp-diffs@lists
Revert Plan:
OK

Blame Revision:
195618