Rework object destruction at shutdown time
Summary:
My previous change on object destruct had a problem and can cause crash
because of destruction order. The ClassStatement can be destroyed first
and the AtomicString name will be freed but the StringData* is still
referenced by the Object. Therefore Object needs to be destroyed first
as the original code did. I made a change so that only calling __destruct
is skipped but the Objects destruction remains as before.

Test Plan:
test/test TestCodeRunEval
run hphpi and browse the site without crash (with TCMALLOC/JEMALLOC off/DEBUG
on)

DiffCamp Revision: 199137
Reviewed By: qixin
Reviewers: hzhao, qixin
CC: qixin, hphp-diffs@lists
Revert Plan:
OK