Work-around GCC bug with vector iterators and optimization levels
Summary:
From hyves

This replaces iterators with indexes for iterating over the parameter
vector in the eval invoke functions.

It works around a bug in GCC where different method signatures are
generated for vector<ExpressionPtr>.begin() (and end()) according to
the optimization level. In one variant a pointer to the place where the
returned iterator must be stored is passed as a parameter to the
function, while in the other variant the iterator is optimized into a
simple pointer and simply put into the return register.

The two different optimization levels come from the fact that the eval
helper functions generated for system are put into
dynamic_table_func.no.cpp and compiled at -O0, while the ones generated
for the PHP codebase (when EnableEval >= 1) are put into the normal cpp
files and compiled at -O3. In the end the linker chooses any of the two
variants of that function when generating the binary (since they share
the same name) thus causing the other half of the function calls to
fail.

This problem is most likely related to the one reported at
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=46143 (and maybe also
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=47011), although it's been
observed in different situations under GCC 4.4.4/5 and 4.5.2. The bug
is also not triggered all the times the function is compiled at
different optimization levels, but only if some unknown additional
conditions hold, which seem to change from one GCC version to the other
anyway.

This workaround modifies the implementation to avoid calling begin()
and end(), thus avoiding the problem altogether.

Test Plan:
make fast_tests
make slow_tests

DiffCamp Revision: 214914
Reviewed By: mwilliams
Reviewers: mwilliams
CC: mwilliams, hphp-diffs@lists
Revert Plan:
ok