Add looping service threads
Summary:
From Hyves:

This adds a new kind of service threads that executes a document in a
loop. It can be useful over normal service threads with an infinite
loop coded in PHP in that memory is collected after each execution of
the document. This allows running a task indefinitely and periodically
recover objects that formed reference cycles by restarting the
execution of the document.

Since the document is executed cyclicly this kind of services is not
fit for initialization tasks that must be executed before the webserver
is started. Therefore they never block the startup of the webserver,
and hphp_service_thread_started is a no-op.

Also wait for service threads to be stopped before shutting down the server,
thus allowing for graceful shutdown.

Test Plan:
make fast_tests
makw slow_tests

DiffCamp Revision: 215534
Reviewed By: qixin
Reviewers: hzhao, qixin, mwilliams
CC: qixin, hphp-diffs@lists
Revert Plan:
Ok