PHP difference: relax method incompatibility check
Summary:
From hyves:

In PHP it is allowed to override a method and set default values for
parameters that didn't have any default value in the base method. HPHPi
instead complains about this.

This reproduces the issue:

Test Plan:
make fast_tests
make slow_test
then:
<?php

interface A {
	public function foo($a);
}

class B implements A {
	public function foo($a = 1) {
		echo $a, PHP_EOL;
	}
}

$b = new B();
$b->foo();

?>

This prints "1" under PHP (and HPHP), but gives a fatal error
"Declaration of B::foo() must be compatible with that of A::foo()" in
HPHPi.

Reviewed By: andrewparoski
Reviewers: andrewparoski, mwilliams
Commenters: mwilliams
CC: andrewparoski, mwilliams
Revert Plan:
Ok

Differential Revision: 225356