Log stacktrace for UncatchableException
Summary:
Infinite recursion and OOM are thrown as UncatchableExceptions, and we have not
been logging their stacktraces, even when Server.InjectedStackTrace is enabled.

This diff adds the stacktrace to the end of the message, and also a runtime
option to limit the maximum depth of the generated stacktrace, in case it is an
infinite recursion.

Test Plan:
Run a simple script in CLI and server mode, with a small memory limit. Saw
stacktraces in OOM messages.

DiffCamp Revision: 201039
Reviewed By: mwilliams
Reviewers: ps, mwilliams
Commenters: ps
CC: ps, qixin, mwilliams, hphp-diffs@lists
Revert Plan:
OK