Vanquish the "ei_" invoke helpers
Summary:
This revision gets rid of the "ei_" invoke helpers and changes hphpi to use
CallInfo instead. To make this work, a new method named "mustBeRef" was
added to CallInfo to allow hphpi to distinguish between arguments that must
be passed by reference (formally declared by-ref arguments and varargs for
functions tagged as ReferenceVariableArgument) and arguments that may be
passed by reference (varargs for functions tagged as MixedVariableArgument).
This distinction is important because it affects whether hphpi throws a
fatal error if a non-refable expression is supplied for an argument that is
pass by reference.

While this change does not significant reduce code size (because we only
generated "ei_" helpers for builtins), it reduces the complexity of our
dynamic invocation system and it gives us better test coverage for the
CallInfo infrastructure.

This change reduces the size of the hphp and hphpi binaries by about 4.3MB
(hphp: 281.8MB -> 277.6MB, hphpi: 247.3MB -> 243.0MB), and it also reduces
the size of www by 4.3MB as well.

Task ID: #

Blame Rev:

Reviewers: qixin, myang, qigao

CC:

Test Plan:
make fast_tests
make slow_tests
Run flib tests under hphpi
Build www and run on local machine

Revert Plan:

Tags: hphp

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 214165