[Perf/fix] Make Variant::set() safe for $a['x'] = $a
Summary:
As in PHP 5.3, an assignment like $a['x'] = $a shouldn't create a recursive
array, no matter whether $a is false/null or it holds the unique reference to an
array.

HPHP has been generating one extra temporary variable for the $a['x'] = $a
case. This diff fixes the problem in the runtime code, so that we can get rid of
the temporary, and also help preventing the problem in the runtime.

Test Plan:
fast_tests
slow_tests

<?php
function f($a) {
  $a[0] = $a;
  var_dump($a);
}
f(false);

<?php
function f($b, $c) {
  $a = $c ? 0 : array('x' => $b);
  $a[0] = $a;
  var_dump($a);
}
f(false, false);

I manually removed temporary variables in the generated code, and it still
behaved correctly.

DiffCamp Revision: 204083
Reviewed By: mwilliams
Reviewers: mwilliams
CC: qixin, mwilliams, hphp-diffs@lists
Tasks:
449093: Temporaries with unnecessary non-ref Variant types

Revert Plan:
OK