[Perf] Generate more direct method calls
Summary:
If a method is resolved to one that is known to be overridden by a same named
method in /any/ other class (and its not perfect virtual), we generate an
indirect call to it. We only need to do that if its overridden by a same named
method in a class derived from the object's class.

This diff makes use of the newly added NotFinal flag on classes to say that if
the object's class is not derived from, we can call the found method directly.

In theory, we could do even better, by marking individual functions as
overridden on a per class basis - but this seems to do pretty well as is.

Test Plan:
fast_tests, slow_tests, flib tests. Inspect diffs for www-dev.

I looked at a number of call-sites, and they appear to be correct, and there
are a lot of changes. Size of www-dev goes down by 1.5M.

DiffCamp Revision: 204834
Reviewed By: myang
Reviewers: qixin, myang
CC: myang, hphp-diffs@lists
Revert Plan:
OK