Turn on ELF buildids for hphp builds
Summary:
The perf tool included in recent kernel revisions refuses to
symbolized stripped binaries unless they have a buildid.

This change turns on buildid only when gold is used as the
linker. When we have upgraded our binutils, we can probably
consider turning it on more widely.

The end result is that the www binary ends up with a
.note.gnu.build-id section.

Test Plan:
Build www
Test that the resulting elf binary has the expected ELF section

DiffCamp Revision: 205574
Reviewed By: qixin
Reviewers: qixin
CC: qixin, lovro, hphp-diffs@lists
Tasks:
475411: Embed ELF build-ids in www binary

Revert Plan:
OK