[Fix] Allow accessing the source root
Summary:
When Server.SafeFileAccess=true, the source root might not be included in the
AllowedDirectories, especially for the sandbox configuration. This diff allows
accessing the source root, so that sandboxes may work even when SafeFileAccess
is on.

Test Plan:
fast_tests
created multiple sandboxes on my devserver, with SafeFileAccess turned on
tested on shared devserver devrs101 with SafeFileAccess turned on

DiffCamp Revision: 206277
Reviewed By: qigao
Reviewers: qigao
Commenters: lklots
CC: lklots, gpatangay, epriestley, arice, qixin, clong, hphp-diffs@lists
Tasks:
479520: 'arc build' fails while loading cache archives

Revert Plan:
OK