[Compile-time] Parallelize preOptimize and postOptimize
Summary:
preOptimize and postOptimize were running single threaded. Add some more
locking and a job scheduler to allow them to be parallelized.

Test Plan:
fast_tests, slow_tests, flib/unit tests.

There are no changes to the generated files for www-dev with inlining turned
off. I made some minor adjustments to the inlining heuristics to stabilize the
results independent of order, which results in some changes to the generated
files with inlining turned on. But the generated files are stable wrt the number
of threads, and if I apply the same changes to the heuristics to trunk, I still
get the same result.

Finally, I added a random_shuffle to the order of the scopes that get fed into
the job scheduler, and still got no differences in the generated code.

I've also turned on my optimizations (minus inlining) by default in
facebook/Makefile, and in facebook/push/Makefile. The builds are still quicker
than before.

DiffCamp Revision: 208616
Reviewed By: qixin
Reviewers: qixin
CC: qixin, hphp-diffs@lists
Revert Plan:
OK