[Fix] Fix --format=lib build
Summary:
The recent header file reorganization caused problems with --format=lib,
because we then generate inline wrapper functions for each method. Because of
the way SmartPtr was implemented, this could require class definitions, rather
than just forward declarations.

In order to fix the dependencies, we would have to move the wrappers out of the
class, so that the class headers could all be included, and then the methods
could be declared.

For now, Im going for the simpler solution of changing the implementation of
SmartPtr so that it doesnt depend on the inheritance hierarchy (in particular,
SmartPtr<T> doesnt need to know that T derives from ObjectData).

Test Plan:
fast_tests, slow_tests, build libcore and www-dev

DiffCamp Revision: 207411
Reviewed By: qixin
Reviewers: qixin, qigao
CC: qixin, hphp-diffs@lists
Revert Plan:
OK