Fix crash in FrameInjection::GetBacktrace()
Summary:
When we throw a exception (e.g., timeout) from the FrameInjection construtor,
we have already finished doCommon() which puts 'this' FrameInjection onto
info->m_top. However, because the construtor isn't finished before the
exception is thrown, the corresponding destructor of 'this' FrameInjection
isn't called. As a result, we have a info->m_top points to an invalid
FrameInjection. If we are lucky (and most of the time we are), the invalid
FrameInjection info->m_top won't be used.

When the exception is thrown, any FrameInjection that already constructed
will be destructed. In ~FrameInjection we invoke check_request_timeout (which
seems to be a misnomer because it also checks other surprises such as OOM, or
even another timeout surprise if the timeout thread has raised the p.timeout
again). If any surprise exists, it calls check_request_surprise which
involves generating a stack trace, or calling FrameInjection::GetBacktrace().
The call will cause SEGV because it will use the invalid info->m_top.

I made a fix to throw the exception before the FrameInjection constructor
does anything useful so the info->m_top won't be updated with an invalid
FrameInjection in case of an exception.

I also added m_popArg to prevent ASSERT failure when the arguments were
not pushed yet due to exception thrown from FrameInjection.

Test Plan:
make fast_tests
make slow_tests
make www and do production test

DiffCamp Revision: 210403
Reviewed By: qigao
Reviewers: qixin, qigao
Commenters: ps
CC: ps, qigao, hphp-diffs@lists
Revert Plan:
OK