[Perf] Use more "few args" functions
Summary:
DynamicFunctionCall and NewObjectExpression never used the "few args" invokers
(and nor did SimpleFunctionCalls that were marked for dynamic invoke), so we
always paid the overhead of creating an array for the arguments, and then
iterating over the array in the dispatcher. Factor out some common code, and use
the "few args" invokers.

Test Plan:
fast_tests, slow_tests, flib/unit tests.

For www-dev, the number of calls to getMeth is down to 294 from 1322, and the
number of calls to getFunc is down to 491 from 42093.
For www, the number of calls to getMeth is down to 179 from 589, and the number
of calls to getFunc is down to 15 from 231.

The huge number of calls to getFunc for www-dev was caused by having lots of
common functions marked as DynamicInvoke. www-dev size reduced by 6Mb, but www
only by 200k

Im seeing a >5% speedup in cache_get_scb, but dont know whether to believe
it...

DiffCamp Revision: 212819
Reviewed By: qixin
Reviewers: qixin, myang
CC: qixin, hphp-diffs@lists
Revert Plan:
OK