[Fix] ReflectionProperty::setAccessible should take an argument
Summary:
ReflectionProperty::setAccessible should take an argument to indicate whether
to set the property to be accessible or not. However, the function is not (and
will not be) implemented, and will throw a ReflectionException if called.

Updated the inconsistencies doc to reflect this (and removed a no-longer
correct inconsistency about "global").

Test Plan:
fast_tests. Call ReflectionProperty::setAccessible.

DiffCamp Revision: 212928
Reviewed By: qixin
Reviewers: qixin
CC: sgrimm, qixin, hphp-diffs@lists
Tasks:

Revert Plan:
OK