Add 'tokenizer' to HPHP's list of supported extensions
Summary:
The get_loaded_extensions() builtin returns an array of the names of the
extensions that the runtime supports. phpsh checks this list when it starts
to make sure the extensions is requires are installed. Unfortunately, there
are several extensions which HPHP supports that it fails to list in the
array returned by get_loaded_extensions().

This revision adds 'tokenizer' to the list, and it fixes the casing for the
PDO extension to be consistent with PHP (PHP uses uppercase, we were using
lowercase).

Test Plan:
make fast_tests
Run a program that dumps the result of get_loaded_extensions(), verify that
  'tokenizer' is there, and verify 'PDO' is uppercase

Differential Revision: 216746
Reviewed By: qixin
Reviewers: qixin
CC: qixin, andrewparoski
Revert Plan:
OK