[Cleanup] Miscellaneous cleanup
Summary:
Moving code around, and adding a number of useful accessor functions.

Test Plan:
fast_tests slow_tests, build www with no change to generated code.

This is the set of changes to the compiler that are needed to support the
bytecode emitter. Im committing to trunk, so I can merge from trunk to hhbc.

DiffCamp Revision: 216850
Reviewed By: andrewparoski
Reviewers: qixin, je, andrewparoski
CC: andrewparoski, mwilliams, hphp-diffs@lists
Revert Plan:
OK