[Fix] SoapClient::__call() was not called from hphpi
Summary:
EvalObjectData needs to check the attribute in addition to the existence
of the method statement of __call(), so that __call() declared in
compiled superclass can be called.

The diff also adds a php_path variable in TestCodeRun to specify the
path to the php binary.

Test Plan:
fast_tests
lbcontrol.php lb13.02.snc4.hphp.nhm

DiffCamp Revision: 217297
Reviewed By: qigao
Reviewers: qigao
CC: qixin, qigao, hphp-diffs@lists
Tasks:
503188: instantDyno lost access to all loadbalancers

Revert Plan:
OK