[Fix/Perf] More accurate hoisting of classes
Summary:
Php hoists some top-level file-scope function and class declarations to the top
of the scope. However, it doesnt hoist a class which has base classes that
werent defined at parse time. This is difficult to model in the compiler, so we
approximate. The previous approximation was dangerous, because it was possible
to construct code that worked in php, but failed at runtime when compiled. This
commit attempts to fix that, so that hphp should be more permissive than php.

I also noticed that we generate a lot of redundant checkClassExists for
volatile classes, so this commit also tries to avoid duplicating these checks.

Test Plan:
fast_tests slow_tests, build www and check that the order of declarations vs
checks is correct for MPathPattern.php. run flib/unit tests against www-dev.

DiffCamp Revision: 218251
Reviewed By: qixin
Reviewers: qixin
CC: qixin, mwilliams, hphp-diffs@lists
Tasks:
502983: Perflab detected a PHP fatal in MPathPattern.php that crashes about
70% of requests

Revert Plan:
OK