[Fix] setArgPop wasnt being called at the right time
Summary:
setArgPop needs to be set to say whether or not the arguments have actually
been pushed.

When calling inovkeImpl, m_argc is set before the args are actually pushed, so
it must be called late. When calling directBind, m_argc is incremented as the
arguments are pushed, so it must be called early. Fix the directBind cases.

Test Plan:
fast_tests, flib tests.

DiffCamp Revision: 218249
Reviewed By: myang
Reviewers: myang
CC: myang, hphp-diffs@lists
Revert Plan:
OK