Better default random number seeder, based on /dev/random
Summary:
When you call rand without seeding or seed without a value we auto-generate a
seed.

The default one is based on some system data, but not really that random.

This uses the openssl RAND_bytes() which reads from /dev/random and if it fails
it goes
back to the old way.

Test Plan:
make
make fast_tests

DiffCamp Revision: 219083
Reviewed By: hzhao
Reviewers: hzhao, mwilliams, lesha
Commenters: amenghra
CC: hzhao, amenghra, macvicar, hphp-diffs@lists
Tasks:
397611: Fwd: Re: [hphp] mt_rand() entropy

Revert Plan:
Ok